requirements.txt
streamlit>=1.28.0
altair>=4.2.0
pandas>=1.3.0
numpy>=1.21.0
requests>=2.28.0
python-dotenv>=0.19.0
msal>=1.18.0
google-adk>=0.1.0
openai==1.91.0
toml>=0.10.2 
